requirements.txt
asgiref==3.8.1
AWSIoTPythonSDK==1.5.4
Django==5.1.1
djangorestframework==3.15.2
paho-mqtt==2.1.0
psycopg2==2.9.9
sqlparse==0.5.1
tzdata==2024.2
